Rename OverlayManager class to WindowManager
