+++ BBTAG_IM.vcxproj
@@ -170,7 +170,7 @@
     <ClCompile Include="src\Web\url_downloader.cpp" />
     <ClCompile Include="src\Overlay\fonts.cpp" />
     <ClCompile Include="src\Overlay\Window\PaletteEditorWindow.cpp" />
-    <ClCompile Include="src\Overlay\OverlayManager.cpp" />
+    <ClCompile Include="src\Overlay\WindowManager.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="depends\imgui\imconfig.h" />
@@ -235,7 +235,7 @@
     <ClInclude Include="src\Web\url_downloader.h" />
     <ClInclude Include="src\Overlay\fonts.h" />
     <ClInclude Include="src\Overlay\Window\PaletteEditorWindow.h" />
-    <ClInclude Include="src\Overlay\OverlayManager.h" />
+    <ClInclude Include="src\Overlay\WindowManager.h" />
   </ItemGroup>
   <ItemGroup>
     <Text Include="resource\donators.txt" />
