+++ src/D3D9EXWrapper/ID3D9EXWrapper_Device.cpp
@@ -5,7 +5,7 @@
 #include "Core/interfaces.h"
 #include "Core/logger.h"
 #include "Hooks/hooks_bbtag.h"
-#include "Overlay/OverlayManager.h"
+#include "Overlay/WindowManager.h"
 
 #include <steam_api.h>
 
@@ -145,11 +145,11 @@ HRESULT APIENTRY Direct3DDevice9ExWrapper::Reset(D3DPRESENT_PARAMETERS* pPresent
 	Settings::applySettingsIni(pPresentationParameters);
 	logD3DPParams(pPresentationParameters, false);
 
-	OverlayManager::getInstance().InvalidateDeviceObjects();
+	WindowManager::getInstance().InvalidateDeviceObjects();
 
 	HRESULT ret = m_Direct3DDevice9Ex->Reset(pPresentationParameters);
 
-	OverlayManager::getInstance().CreateDeviceObjects();
+	WindowManager::getInstance().CreateDeviceObjects();
 
 	return ret;
 }
@@ -345,7 +345,7 @@ HRESULT APIENTRY Direct3DDevice9ExWrapper::BeginScene()
 {
 	LOG(7, "BeginScene\n");
 
-	OverlayManager::getInstance().Update();
+	WindowManager::getInstance().Update();
 
 	return m_Direct3DDevice9Ex->BeginScene();
 }
@@ -354,7 +354,7 @@ HRESULT APIENTRY Direct3DDevice9ExWrapper::EndScene()
 {
 	LOG(7, "EndScene\n");
 
-	OverlayManager::getInstance().Render();
+	WindowManager::getInstance().Render();
 
 	return m_Direct3DDevice9Ex->EndScene();
 }
@@ -966,11 +966,11 @@ HRESULT APIENTRY Direct3DDevice9ExWrapper::ResetEx(D3DPRESENT_PARAMETERS* pPrese
 	Settings::applySettingsIni(pPresentationParameters);
 	logD3DPParams(pPresentationParameters, false);
 	
-	OverlayManager::getInstance().InvalidateDeviceObjects();
+	WindowManager::getInstance().InvalidateDeviceObjects();
 
 	HRESULT ret = m_Direct3DDevice9Ex->ResetEx(pPresentationParameters, pFullscreenDisplayMode);
 
-	OverlayManager::getInstance().CreateDeviceObjects();
+	WindowManager::getInstance().CreateDeviceObjects();
 
 	return ret;
 }
