+++ src/SteamApiWrapper/SteamApiHelper.cpp
@@ -1,7 +1,7 @@
 #include "SteamApiHelper.h"
 
 #include "Core/logger.h"
-#include "Overlay/OverlayManager.h"
+#include "Overlay/WindowManager.h"
 
 //note: commented out functions are not working
 
@@ -26,13 +26,13 @@ void SteamApiHelper::OnUpdateNumberOfCurrentPlayers(NumberOfCurrentPlayers_t *pC
 {
 	if (bIOFailure || !pCallback->m_bSuccess)
 	{
-		OverlayManager::getInstance().AddLog("[error] Failed to update the current number of ingame players\n");
+		WindowManager::getInstance().AddLog("[error] Failed to update the current number of ingame players\n");
 		current_players = -1;
 		return;
 	}
 	//printf("Number of players currently playing: %d\n", pCallback->m_cPlayers);
 #ifdef ENABLE_LOGGING
-	OverlayManager::getInstance().AddLog("[debug] Updated the current number of ingame players: %d\n", pCallback->m_cPlayers);
+	WindowManager::getInstance().AddLog("[debug] Updated the current number of ingame players: %d\n", pCallback->m_cPlayers);
 #endif
 	current_players = pCallback->m_cPlayers;
 }
