+++ src/Web/url_downloader.cpp
@@ -1,7 +1,7 @@
 #include "url_downloader.h"
 
 #include "Core/utils.h"
-#include "Overlay/OverlayManager.h"
+#include "Overlay/WindowManager.h"
 
 #include <wininet.h>
 
@@ -15,7 +15,7 @@ std::string DownloadUrl(std::wstring& wUrl)
 
 	if (!connect)
 	{
-		OverlayManager::getInstance().AddLog("[error] DownloadUrl failed. Connection Failed or Syntax error with URL\n'%s'\n", url.c_str());
+		WindowManager::getInstance().AddLog("[error] DownloadUrl failed. Connection Failed or Syntax error with URL\n'%s'\n", url.c_str());
 		return "";
 	}
 
@@ -24,7 +24,7 @@ std::string DownloadUrl(std::wstring& wUrl)
 	if (!OpenAddress)
 	{
 		DWORD ErrorNum = GetLastError();
-		OverlayManager::getInstance().AddLog("[error] DownloadUrl failed. Failed to open URL\n'%s'\ncode: %d\n", url.c_str(), ErrorNum);
+		WindowManager::getInstance().AddLog("[error] DownloadUrl failed. Failed to open URL\n'%s'\ncode: %d\n", url.c_str(), ErrorNum);
 		InternetCloseHandle(connect);
 		return "";
 	}
@@ -52,7 +52,7 @@ unsigned long DownloadUrlBinary(std::wstring& wUrl, void** outBuffer)
 
 	if (!connect)
 	{
-		OverlayManager::getInstance().AddLog(
+		WindowManager::getInstance().AddLog(
 			"[error] DownloadUrlBinary failed. Connection Failed or Syntax error with URL\n'%s'\n",
 			url.c_str()
 		);
@@ -64,7 +64,7 @@ unsigned long DownloadUrlBinary(std::wstring& wUrl, void** outBuffer)
 	if (!OpenAddress)
 	{
 		DWORD ErrorNum = GetLastError();
-		OverlayManager::getInstance().AddLog(
+		WindowManager::getInstance().AddLog(
 			"[error] DownloadUrlBinary failed. Failed to open URL\n'%s'\ncode: %d\n",
 			url.c_str(), ErrorNum
 		);
